+++ src/zenml/zen_stores/schemas/user_schemas.py
@@ -44,7 +44,6 @@ if TYPE_CHECKING:
         EventSourceSchema,
         FlavorSchema,
         ModelSchema,
-        ModelVersionArtifactSchema,
         ModelVersionPipelineRunSchema,
         ModelVersionSchema,
         OAuthDeviceSchema,
@@ -143,9 +142,6 @@ class UserSchema(NamedSchema, table=True):
     )
     model_versions: List["ModelVersionSchema"] = Relationship(
         back_populates="user",
-    )
-    model_versions_artifacts_links: List["ModelVersionArtifactSchema"] = (
-        Relationship(back_populates="user")
     )
     model_versions_pipeline_runs_links: List[
         "ModelVersionPipelineRunSchema"
